services/voting/requirements_prod.txt
flask
flask_httpauth
sqlalchemy
sqlalchemy-utils
sqlalchemy-bigquery
google-cloud-pubsub